[FIX] sale_timesheet_ux: Generate a project from template.
Change condition before generate project from template: formerly asking if
there was not analytic account, now we ask if there is not projects related to
the analytic account.

This is required because the analytic account could be created and related to
the sale order before confirmation, in this case no project is created and
neither the generate from project functionality since the created analytic
account condition.