[REF] sale_order_type_invoice_policy: Readapt funcionality
In this case we add new invoice policy type "prepaid_block_delivery" on sale order type to block the delivery validate if there aren't sale order lines invoiced and paid to the related sale order of the picking.
The policy "prepaid" now are rename as "Prepaid - Block Reserve" to able to validate and not to reserve quanty of the product for the same case of the previus policy.

This way you can validate the sale order and depends the policy chosen on the sale order type we restring to reserve or validate the picking related.