[13.0] [FIX] sale_order_type_user_default: Save the value that you setting in.
Before that changes the field for user default value for the user is saving , although you select a different value after this was is setting.
We need to only show this value (user default) when create and if you change it when save the record the value keep the same that the user choose, not the default value.