[FIX] sale_order_type_ux: Fix tax computation when invoice company differs from sale order company and having fiscal position
Overrides the `_create_invoices` method to ensure taxes are correctly computed
based on the company of the generated invoice, not the sale order. This is
important in multi-company environments or when localizations like l10n_ar
are used, as taxes may differ by company taking in account the fiscal position.

closes ingadhoc/sale#1234

Signed-off-by: Camila Vives <cav@adhoc.com.ar>