[FIX] sale_order_type_ux: browse fiscal position from second company
The original code checked if the fiscal position's company was different from the sales order's company. If the second company (the invoice company) did not have a fiscal position, nothing was assigned.

I added a condition so that if no fiscal position is defined, it will always look for the fiscal position of the invoice company.

closes ingadhoc/sale#1287

Signed-off-by: rov-adhoc <rov@adhoc.com.ar>